Extracting Security to a separate file.
